+++ googledrive/api.py
@@ -31,7 +31,8 @@ class GoogleAuth:
                 creds.refresh(Request())
             else:
                 flow = InstalledAppFlow.from_client_secrets_file(
-                    credentials, GoogleAuth.SCOPES)
+                    credentials,
+                    scopes)
                 creds = flow.run_local_server(port=0)
             # Save the credentials for the next run
             with open('token.pickle', 'wb') as token: