+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## 0.1.3
+
+- fixing cli usecases
+  - google-drive ls
+  - google-drive mkdir
+- adding usage of click for cli interaction
+
 ## 0.1.2
 
 - fixing api.GoogleService.__get_credentials function, adding scopes argument to its __init__ funciton